backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
tensorflow>=2.15.0
Pillow>=10.0.0
numpy>=1.24.0
reportlab>=4.1.0
